+++ src/training/scheduler.h
@@ -10,7 +10,6 @@ class Scheduler : public TrainingObserver {
 private:
   Ptr<Config> options_;
   std::vector<Ptr<ValidatorBase>> validators_;
-  bool validated_{false};
 
   bool first_{true};
 
@@ -44,10 +43,7 @@ public:
 
   void increaseEpoch() {
     LOG(info, "Seen {} samples", state_->samples);
-
     state_->newEpoch();
-    state_->samples = 0;
-
     LOG(info, "Starting epoch {}", state_->epochs);
   }
 
@@ -57,8 +53,13 @@ public:
   void addValidator(Ptr<ValidatorBase> validator) {
     validators_.push_back(validator);
     // stalled validations are computed with the first validator only
-    if(validators_.size() == 1)
+    if(validators_.size() == 1) {
       registerTrainingObserver(validators_.front());
+      // initialize the value of the very first validation score, which should
+      // be the worst possible
+      if(!state_->loaded)
+        state_->validBest = validators_.front()->initScore();
+    }
   }
 
   bool validating() {
@@ -71,8 +72,9 @@ public:
   }
 
   void validate(const std::vector<Ptr<ExpressionGraph>>& graphs, bool final = false) {
-    if(validated_ || (state_->batches % options_->get<size_t>("valid-freq") != 0
-                      && !final))
+    if(state_->validated
+       || (state_->batches % options_->get<size_t>("valid-freq") != 0
+           && !final))
       return;
 
     bool firstValidator = true;
@@ -106,7 +108,7 @@ public:
       firstValidator = false;
     }
 
-    validated_ = true;
+    state_->validated = true;
   }
 
   size_t stalled() {
@@ -117,6 +119,8 @@ public:
   }
 
   void update(float cost, Ptr<data::Batch> batch) {
+    state_->validated = false;
+
     state_->costSum += cost * batch->size();
     state_->samples += batch->size();
     state_->samplesDisp += batch->size();
@@ -151,25 +155,30 @@ public:
       state_->wordsDisp = 0;
       state_->samplesDisp = 0;
     }
-
-    validated_ = false;
   }
 
   void load(const std::string& name) {
-    std::string nameYaml = name + ".yml";
-    if(boost::filesystem::exists(nameYaml)) {
-      YAML::Node config = YAML::LoadFile(nameYaml);
-      state_->load(config);
+    std::string nameYaml = name + ".progress.yml";
+    if(boost::filesystem::exists(nameYaml))
+      state_->load(nameYaml);
+
+    // @TODO: not sure if needed
+    if(!options_->get<bool>("restore-corpus")) {
+      state_->costSum = 0;
+      state_->samplesDisp = 0;
+      state_->wordsDisp = 0;
     }
+
     state_->newLoad();
   }
 
   void save(const std::string& name) {
+    // Save config options
     YAML::Node config = options_->get();
-    state_->save(config);
-    std::string nameYaml = name + ".yml";
-    std::ofstream fout(nameYaml);
+    std::ofstream fout(name + ".yml");
     fout << config;
+    // Save training progress
+    state_->save(name + ".progress.yml");
   }
 
   size_t numberOfBatches() { return state_->batches; }