+++ src/training/training.h
@@ -47,20 +47,29 @@ public:
         scheduler->addValidator(validator);
     }
 
+    auto batchGenerator = New<CorpusBatchGenerator>(dataset, options_, stats);
+    scheduler->registerTrainingObserver(batchGenerator);
+
     auto model = New<ModelWrapper>(options_);
     model->setScheduler(scheduler);
     model->load();
 
-    auto batchGenerator = New<BatchGenerator<CorpusBase>>(dataset, options_, stats);
+    // @TODO: shuffle_ as a private attribute in BG
+    auto shuffle = !options_->get<bool>("no-shuffle");
+    bool restored = options_->get<bool>("restore-corpus")
+                    && batchGenerator->restore(trainState, shuffle);
 
     scheduler->started();
     while(scheduler->keepGoing()) {
-      auto shuffle = !options_->get<bool>("no-shuffle");
-      batchGenerator->prepare(shuffle);
+      if(!restored)
+        batchGenerator->prepare(shuffle);
+      restored = false;
+
       while(*batchGenerator && scheduler->keepGoing()) {
         auto batch = batchGenerator->next();
         model->update(batch);
       }
+
       if(scheduler->keepGoing())
         scheduler->increaseEpoch();
     }