+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Restoring optimizer and training parameters such as learning rate, validation
   results, etc.
 - Experimental CPU training/translation with `--cpu-threads=N`
+- Restoring corpus iteration after training is restarted
 
 ### Fixed
 - Deterministic data shuffling with specific seed for SQLite3 corpus storage
@@ -25,8 +26,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [1.3.1] - 2018-02-04
 
 ### Fixed
-- Fixed missing final validation when done with training
-- Fixed differing summaries for marian-scorer when used with multiple GPUs
+- Missing final validation when done with training
+- Differing summaries for marian-scorer when used with multiple GPUs
 
 ## [1.3.0] - 2018-01-24